:recycle: move the widget class to namespace
and leave behind a stub class that just inherits everything from the new namespaced class
